requirements.txt
# DCRM Analyzer Pro - Master Requirements
# Generated from codebase scan

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
openpyxl>=3.1.0  # Required for pandas read_excel

# Web Frameworks & APIs
streamlit>=1.30.0
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0

# Google AI / LangChain
langchain-google-genai>=2.0.0
langchain-experimental>=0.3.0
langchain-core>=0.3.0
google-generativeai>=0.8.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning / PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Database
pymongo>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
